stopService removes its entries from the network address cache (#9)
* stopService removes its entries from the network address cache

* fix: kill tailed logs child hangs if already exited/killed previously

* add tests for cache usage of getAddressForService

+++ src/container-logs.ts
@@ -35,6 +35,7 @@ export function tailLogsForServices(
   }
 
   return async () => {
+    if (child.killed || child.exitCode !== null) return
     child.kill(9)
     await retry(() => {
       if (!child.killed) {
